Move bets back into single file bets.py
