--- crapssim/bet/__init__.py
@@ -1,7 +1,0 @@
-from crapssim.bet.base import *
-from crapssim.bet.hard_way import *
-from crapssim.bet.pass_line import *
-from crapssim.bet.odds import *
-from crapssim.bet.one_roll import *
-from crapssim.bet.place import *
-from crapssim.bet.side import *
