+++ tests/integration/test_bet.py
@@ -1,12 +1,9 @@
 import pytest
 
-from crapssim import Table, Dice
-from crapssim.bet import PassLine, Come, Odds
-from crapssim.bet.hard_way import HardWay
-from crapssim.bet.one_roll import Field, Any7, Two, Three, Yo, Boxcars, AnyCraps, CAndE
-from crapssim.bet.pass_line import DontPass, DontCome
-from crapssim.bet.place import Place
-from crapssim.bet.side import Fire
+from crapssim import Dice, Table
+from crapssim.bet import (Any7, AnyCraps, Boxcars, CAndE, Come, DontCome,
+                          DontPass, Field, Fire, HardWay, Odds, PassLine,
+                          Place, Three, Two, Yo)
 from crapssim.point import Point
 from crapssim.table import TableUpdate
