+++ tests/integration/test_layodds_integration.py
@@ -1,7 +1,6 @@
 import pytest
 
-from crapssim.bet import Odds
-from crapssim.bet.pass_line import DontPass
+from crapssim.bet import DontPass, Odds
 from crapssim.strategy import BetDontPass, DontPassOddsMultiplier
 from crapssim.table import Table, TableUpdate
